--- src/main/java/no/entur/uttu/export/messaging/MessagingService.java
@@ -1,5 +1,0 @@
-package no.entur.uttu.export.messaging;
-
-public interface MessagingService {
-  void notifyExport(String codespace, String filename);
-}
