+++ src/main/java/no/entur/uttu/export/messaging/NoopMessagingService.java
@@ -1,8 +1,12 @@
 package no.entur.uttu.export.messaging;
 
+import no.entur.uttu.export.messaging.spi.MessagingService;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.stereotype.Component;
 
+/**
+ * The default implementation is a noop operation
+ */
 @Component
 @ConditionalOnMissingBean(
         value = MessagingService.class, ignored = NoopMessagingService.class
@@ -10,6 +14,6 @@ import org.springframework.stereotype.Component;
 public class NoopMessagingService implements MessagingService {
     @Override
     public void notifyExport(String codespace, String filename) {
-        // noop
+        // intentionally empty
     }
 }
